+++ changelog.md
@@ -1,3 +1,9 @@
+## 0.2.14 TBA
+
+ - **feat**: add an Observable interface to watchers
+ - **feat**: include the original watch response as the third argument in watch put/delete events.
+ - **fix**: fixed wrong data type annotation in `watcher.lastRevision()`.
+
 ## 0.2.13 2019-07-03
 
  - **bug**: fixed comparisons breaking in STM when using namespaces (see [#90](https://github.com/mixer/etcd3/issues/90))